+++ src/zukan_icon_theme/utils/contexts_scopes.py
@@ -58,6 +58,7 @@ CONTEXTS_SCOPES = [
     {'scope': 'text.html.basic', 'startsWith': 'XML (SVG)'},
     {'scope': 'text.tex.latex', 'startsWith': 'Literate Haskell'},
     # Yaml Pipelines
+    {'scope': 'source.dockerfile', 'startsWith': 'Dockerfile (build)'},
     {'scope': 'source.yaml.helm', 'startsWith': 'YAML (Helm)'},
     {'scope': 'source.yaml.kong', 'startsWith': 'YAML (Kong)'},
     {'scope': 'source.yaml.pipeline.bitbucket', 'startsWith': 'YAML (Bitbucket)'},